Remove int cast after changing to size_t

+++ swaylock/main.c
@@ -716,7 +716,7 @@ static char *get_config_path(void) {
 
 	wordexp_t p;
 	char *path;
-	for (size_t i = 0; i < (int)(sizeof(config_paths) / sizeof(char *)); ++i) {
+	for (size_t i = 0; i < (sizeof(config_paths) / sizeof(char *)); ++i) {
 		if (wordexp(config_paths[i], &p, 0) == 0) {
 			path = strdup(p.we_wordv[0]);
 			wordfree(&p);
